h3: make all events edge-triggered
Currently all HTTP/3 events are edge-triggered with the exception of
`Data` and `Datagram`, which get reported until the underlying data is
read by the application.

Besides being inconsistent, this is problematic because an application
can't decide to ignore those events in particular circumstances. For
example an application might want to apply backpressure to a particular
stream on which the Data event has been reported, but still process
other streams normally.

In the current implementation, a Data event for the "blocked" stream
would keep being fired until the application decides to actually read
the body data from that stream, which means that, if the application is
not ready to do that yet, no other events are reported, thus blocking
all other streams as well.

This changes those events to only be triggered once, until re-armed.

Data events for a particular stream are re-armed after all queued data
for that stream is read by the application (including data from multiple
DATA frames already buffered).

Datagram events are re-armed once all queued DATAGRAMs for a specific
flow ID are read by the application (that is, if multiple DATAGRAMs are
queued next to each other, only one Datagram event for that flow ID is
fired), or when the incoming DATAGRAM queue is empty.

This also makes the poll count thresholds superfluous as each event is
only reported once rather than forever.

Note that applications will now need to make sure to read all of the
data from a stream, otherwise they might get stuck (see examples and
apps updates), so this is a breaking change.